==> notebooks/01_load_and_validate_raw_data.ipynb <==
import numpy as np
import pandas as pd

from pathlib import Path
import requests

def download_one_file_of_raw_data(year:int,month:int) -> Path:

    URL=f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"

    response = requests.get(URL)

    if response.status_code ==200:
        path = f'../data/raw/rides_{year}-{month:02d}.parquet'
        open(path,"wb").write(response.content)
        return path
    else:
        raise Exception(f"{URL} is not available")

cd taxi_demand_predictor


cd notebooks

download_one_file_of_raw_data(year=2022,month=1)

rides = pd.read_parquet("../data/raw/rides_2022-01.parquet")
rides.head(20)

rides = rides[['tpep_pickup_datetime','PULocationID']]
rides.rename(columns={'tpep_pickup_datetime': 'pickup_datetime','PULocationID':'pickup_location_id'},inplace=True)

rides.head(20)

rides["pickup_datetime"].describe()

rides = rides[rides.pickup_datetime >='2022-01-01']
rides = rides[rides.pickup_datetime < '2022-02-01']

rides['pickup_datetime'].describe()

rides.to_parquet("../data/transformed/validated_rides_2022_01.parquet")
